+++ lamin_cli/_rest/__init__.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
 from ._click import rest_group
-from ._client import request_json
 from ._mutations import delete, insert, update, upsert
 from ._query import get_record, list_records
 from ._schema import schema
 from ._statistics import relation_counts, statistics
+from ._utils import request_json
 
 
 @rest_group
